More file reorg, add tab-view
