+++ PalCalc.UI/View/Inspector/SaveInspectorWindow.xaml
@@ -3,117 +3,19 @@
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:PalCalc.UI.View"
+        xmlns:local="clr-namespace:PalCalc.UI.View.Inspector"
         xmlns:svm="clr-namespace:PalCalc.UI.ViewModel.Inspector"
         mc:Ignorable="d"
         WindowStyle="ToolWindow"
         WindowStartupLocation="CenterOwner"
-        d:DataContext="{x:Static svm:SaveInspectorViewModel.DesignerInstance}"
+        d:DataContext="{x:Static svm:SaveInspectorWindowViewModel.DesignerInstance}"
         Title="Save File Inspector" Height="450" Width="900">
-    <Grid>
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="149*" MaxWidth="300" />
-            <ColumnDefinition Width="Auto" />
-            <ColumnDefinition Width="651*" />
-            <ColumnDefinition Width="Auto" />
-            <ColumnDefinition Width="300*" />
-        </Grid.ColumnDefinitions>
-        <ListBox Grid.Column="0" ItemsSource="{Binding Containers}" SelectedItem="{Binding SelectedContainer}" IsSynchronizedWithCurrentItem="True">
-            <ListBox.ItemTemplate>
-                <DataTemplate DataType="svm:InspectedContainerViewModel">
-                    <StackPanel>
-                        <TextBlock>
-                            <Run Text="Type: " />
-                            <Run Text="{Binding Type, Mode=OneTime}" />
-                        </TextBlock>
-                        <TextBlock>
-                            <Run Text="Owner: " />
-                            <Run Text="{Binding Owner.Name, Mode=OneTime}" />
-                        </TextBlock>
-                        <TextBlock>
-                            <Run Text="{Binding UsedSlots, Mode=OneTime}" />
-                            <Run>/</Run>
-                            <Run Text="{Binding TotalSlots, Mode=OneTime}" />
-                            <Run>slots used</Run>
-                        </TextBlock>
-                        <TextBlock>
-                            <Run>ID:</Run>
-                            <Run Text="{Binding Id, Mode=OneTime}" />
-                        </TextBlock>
-                    </StackPanel>
-                </DataTemplate>
-            </ListBox.ItemTemplate>
-        </ListBox>
-
-        <GridSplitter Grid.Column="1" VerticalAlignment="Stretch" HorizontalAlignment="Center" Width="5" />
-
-        <ListView ItemsSource="{Binding SelectedContainer.Slots}"
-                  Grid.Column="2"
-                  SelectedItem="{Binding SelectedSlot}"
-                  IsSynchronizedWithCurrentItem="True"
-                  ScrollViewer.HorizontalScrollBarVisibility="Disabled"
-                  >
-            <ListView.ItemsPanel>
-                <ItemsPanelTemplate>
-                    <WrapPanel />
-                </ItemsPanelTemplate>
-            </ListView.ItemsPanel>
-            <ListView.ItemTemplate>
-                <DataTemplate>
-                    <ContentControl Content="{Binding}" Margin="5">
-                        <ContentControl.Resources>
-                            <DataTemplate DataType="{x:Type svm:EmptyPalContainerSlotViewModel}">
-                                <TextBlock HorizontalAlignment="Center" VerticalAlignment="Center">Empty</TextBlock>
-                            </DataTemplate>
-                            <DataTemplate DataType="{x:Type svm:PalContainerSlotViewModel}">
-                                <StackPanel Orientation="Vertical">
-                                    <Image Source="{Binding Icon}" Width="50" Height="50" Margin="0,0,5,0" />
-                                    <TextBlock Text="{Binding DisplayName}" HorizontalAlignment="Center" />
-                                </StackPanel>
-                            </DataTemplate>
-                        </ContentControl.Resources>
-                    </ContentControl>
-                </DataTemplate>
-            </ListView.ItemTemplate>
-        </ListView>
-
-        <GridSplitter Grid.Column="3" VerticalAlignment="Stretch" HorizontalAlignment="Center" Width="5" />
-
-        <Grid Grid.Column="4">
-            <Grid.RowDefinitions>
-                <RowDefinition Height="Auto" />
-                <RowDefinition Height="Auto" />
-                <RowDefinition Height="1*" />
-            </Grid.RowDefinitions>
-            <GroupBox Header="Detected Properties" Grid.Row="0">
-                <DataGrid AutoGenerateColumns="False"
-                          HeadersVisibility="None"
-                          ItemsSource="{Binding SelectedSlotDetails.PalProperties}"
-                          IsReadOnly="True"
-                          VerticalScrollBarVisibility="Visible"
-                          >
-                    <DataGrid.Columns>
-                        <DataGridTextColumn Width="50*" Binding="{Binding Key}" />
-                        <DataGridTextColumn Width="50*" Binding="{Binding Value}" />
-                    </DataGrid.Columns>
-                </DataGrid>
-            </GroupBox>
-
-            <GridSplitter HorizontalAlignment="Stretch" Grid.Row="1" Height="5" />
-
-            <GroupBox Header="Raw Properties" Grid.Row="2">
-                <DataGrid AutoGenerateColumns="False"
-                          HeadersVisibility="None"
-                          ItemsSource="{Binding SelectedSlotDetails.RawProperties}"
-                          IsReadOnly="True"
-                          VerticalScrollBarVisibility="Visible"
-                          >
-                    <DataGrid.Columns>
-                        <DataGridTextColumn Width="50*" Binding="{Binding Key}" />
-                        <DataGridTextColumn Width="50*" Binding="{Binding Value}" />
-                    </DataGrid.Columns>
-                </DataGrid>
-            </GroupBox>
-        </Grid>
-    </Grid>
+    <TabControl>
+        <TabItem Header="Search">
+            
+        </TabItem>
+        <TabItem Header="Details">
+            <local:SaveDetailsView DataContext="{Binding Details}" />
+        </TabItem>
+    </TabControl>
 </Window>
